P-11 🟠 Ledger and payouts: vendor, admin, marketer and shipping amounts reconciliation
Marketer/agent payout jobs are deferred (orders has no marketer attribution
columns yet — that's P-12's job to add). Everything else: double-pay fix,
extended account_type enum, storage/overage/packaging/subscription
deductions, scheduled vendor payouts per cadence, and financial-report vs.
ledger reconciliation are done and tested.

+++ backend/app/Models/Vendor.php
@@ -82,7 +82,7 @@ class Vendor extends Model
             'last_login_at' => 'datetime',
             'payout_hold_active' => 'boolean',
             'store_rating_avg' => 'float',
-            'total_sales' => 'float',
+            'total_sales' => 'integer',
             'return_rate_pct' => 'float',
             'cancellation_rate_pct' => 'float',
             'strikes_count' => 'integer',
